itemController
доделать

+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,7 +1,6 @@
 package ru.practicum.shareit.user.model;
 
-import lombok.Data;
-import lombok.NonNull;
+import lombok.*;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.LinkedList;
@@ -11,11 +10,10 @@ import java.util.TreeSet;
 /**
  * TODO Sprint add-controllers.
  */
-
 @Data
 public class User {
 
-    private int id;
+    private Long id;
     private String name;
     private String email;
